rename barrel file and add item names for helm barrels
