+++ randomizer/Lists/Item.py
@@ -83,16 +83,16 @@ ItemList = {
     Items.CrystalCavesKey: Item("Crystal Caves Key", True),
     Items.CreepyCastleKey: Item("Creepy Castle Key", True),
     Items.HideoutHelmKey: Item("Hideout Helm Key", True),
-    Items.HelmDonkey1: Item("", False),
-    Items.HelmDonkey2: Item("", False),
-    Items.HelmDiddy1: Item("", False),
-    Items.HelmDiddy2: Item("", False),
-    Items.HelmLanky1: Item("", False),
-    Items.HelmLanky2: Item("", False),
-    Items.HelmTiny1: Item("", False),
-    Items.HelmTiny2: Item("", False),
-    Items.HelmChunky1: Item("", False),
-    Items.HelmChunky2: Item("", False),
+    Items.HelmDonkey1: Item("Helm Donkey Barrel 1", False),
+    Items.HelmDonkey2: Item("Helm Donkey Barrel 2", False),
+    Items.HelmDiddy1: Item("Helm Diddy Barrel 1", False),
+    Items.HelmDiddy2: Item("Helm Diddy Barrel 2", False),
+    Items.HelmLanky1: Item("Helm Lanky Barrel 1", False),
+    Items.HelmLanky2: Item("Helm Lanky Barrel 2", False),
+    Items.HelmTiny1: Item("Helm Tiny Barrel 1", False),
+    Items.HelmTiny2: Item("Helm Tiny Barrel 2", False),
+    Items.HelmChunky1: Item("Helm Chunky Barrel 1", False),
+    Items.HelmChunky2: Item("Helm Chunky Barrel 2", False),
     Items.GoldenBanana: Item("Golden Banana", False),
     Items.BananaFairy: Item("Banana Fairy", False),
     Items.BananaMedal: Item("Banana Medal", False),
